// src/js/models/cim-dictionary/CimDictionaryMeasurement.ts
export interface CimDictionaryMeasurement {
  name: string;
  mRID: string;
  ConductingEquipment_mRID: string;
  Terminal_mRID: string;
  measurementType: string;
  phases: string;
  MeasurementClass: string;
  ConductingEquipment_type: string;
  ConductingEquipment_name: string;
  ConnectivityNode: string;
}